+++ Sources/PositronicKit/PositronicKit+Dependencies.swift
@@ -14,7 +14,7 @@ import PKUtilities
 ///
 /// Not part of the public API surface.
 internal struct KitDependencies: Sendable {
-    var languageModel: any LanguageModel
+    var languageModel: any LLMStreamClient & LLMUtilityClient
     var messageStore: any ThreadMessageStoreProtocol
     var agentInstanceStore: any AgentInstanceStoreProtocol
     var requestOriginStore: any RequestOriginStoreProtocol